refactor(repl): move repl stream into a separate file
